refactor(server): isolate password authentication routes
